Datatype

Description

This module contains all the custom datatypes used all across the applciation.

User and MsgStruct are datatypes used to define the structure of the User who will use the application with message structure that will be sent.

Written by Vishal

Documentation

data User

Datatype for User Structure

Constructors

User

Fields

name :: String

data MsgStruct

Datatype for Message Structure

Constructors

MsgStruct

Fields

sender :: User

receiver :: User

message :: String

users :: [User]

List of Users