Functions

Description

This module contains all the methods used within the application including the methods to manipulate threads, create-send messages. Written by Vishal

Documentation

selectUser :: Int -> IO User

Method to select user from the List of Users

sendMessageThreads :: User -> [User] -> MVar Int -> MVar [MsgStruct] -> IO ()

Method to mimic sending of message from Sender to Receiver using threads. 100 message threads are initiated at random intervals

countMessages :: [MsgStruct] -> Map User Int

Method to find top 3 most active users (Users who received most messages)